LayoutTests/platform/gtk/inspector/canvas/recording-bitmaprenderer-expected.txt

Test that CanvasManager is able to record actions made to BitmapRenderer canvas contexts.


== Running test suite: Canvas.recordingBitmapRenderer
-- Running test case: Canvas.recordingBitmapRenderer.singleFrame
initialState:
  attributes:
    width: 2
    height: 2
  parameters:
  content: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAIAAAACCAYAAABytg0kAAAABHNCSVQICAgIfAhkiAAAAAtJREFUCJljYEAHAAASAAFv+mO3AAAAAElFTkSuQmCC"
frames:
  0: (duration)
    0: transferFromImageBitmap([object ImageBitmap])
      swizzleTypes: [ImageBitmap]
      trace:
        0: transferFromImageBitmap
        1: (anonymous function)
        2: executeFrameFunction
        3: (anonymous function)
        4: asyncFunctionResume
        5: (anonymous function)
        6: promiseReactionJob
      snapshot: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAIAAAACCAYAAABytg0kAAAABHNCSVQICAgIfAhkiAAAABRJREFUCJlj/M/A8J+BgYGBiQEKAB8XAgK+3Bg8AAAAAElFTkSuQmCC"

-- Running test case: Canvas.recordingBitmapRenderer.multipleFrames
initialState:
  attributes:
    width: 2
    height: 2
  parameters:
  content: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAIAAAACCAYAAABytg0kAAAABHNCSVQICAgIfAhkiAAAAAtJREFUCJljYEAHAAASAAFv+mO3AAAAAElFTkSuQmCC"
frames:
  0: (duration)
    0: transferFromImageBitmap([object ImageBitmap])
      swizzleTypes: [ImageBitmap]
      trace:
        0: transferFromImageBitmap
        1: (anonymous function)
        2: executeFrameFunction
        3: (anonymous function)
        4: asyncFunctionResume
        5: (anonymous function)
        6: promiseReactionJob
      snapshot: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAIAAAACCAYAAABytg0kAAAABHNCSVQICAgIfAhkiAAAABRJREFUCJlj/M/A8J+BgYGBiQEKAB8XAgK+3Bg8AAAAAElFTkSuQmCC"

-- Running test case: Canvas.recordingBitmapRenderer.memoryLimit
initialState:
  attributes:
    width: 2
    height: 2
  parameters:
  content: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAIAAAACCAYAAABytg0kAAAABHNCSVQICAgIfAhkiAAAAAtJREFUCJljYEAHAAASAAFv+mO3AAAAAElFTkSuQmCC"
frames:
  0: (duration) (incomplete)
    0: transferFromImageBitmap([object ImageBitmap])
      swizzleTypes: [ImageBitmap]
      trace:
        0: transferFromImageBitmap
        1: (anonymous function)
        2: executeFrameFunction
        3: (anonymous function)
        4: asyncFunctionResume
        5: (anonymous function)
        6: promiseReactionJob
      snapshot: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAIAAAACCAYAAABytg0kAAAABHNCSVQICAgIfAhkiAAAAAtJREFUCJljYEAHAAASAAFv+mO3AAAAAElFTkSuQmCC"

-- Running test case: Canvas.recordingBitmapRenderer.Console
PASS: The recording should have the name "TEST".
PASS: The recording should have one frame.
PASS: The first frame should have one action.
